# Conflicts:
#	.travis.yml
#	1_reloader.py
#	2_bootstrap.py
#	Default.sublime-commands
#	LICENSE
#	Main.sublime-menu
#	PackagesManager.sublime-settings
#	package-metadata.json
#	package_control/automatic_upgrader.py
#	package_control/bootstrap.py
#	package_control/ca_certs.py
#	package_control/commands/__init__.py
#	package_control/commands/existing_packages_command.py
#	package_control/commands/package_control_tests_command.py
#	package_control/commands/remove_package_command.py
#	package_control/commands/upgrade_all_packages_command.py
#	package_control/download_manager.py
#	package_control/http_cache.py
#	package_control/loader.py
#	package_control/tests/__init__.py
#	package_control/tests/all_tests/clients_tests.py
#	package_control/tests/all_tests/providers_tests.py

+++ package_control/loader.py
@@ -6,6 +6,7 @@ from os import path
 import zipfile
 import zipimport
 import shutil
+import time
 from textwrap import dedent
 from threading import Event, Lock
 
@@ -31,10 +32,10 @@ class SwapEvent():
         return not self._ev.is_set()
 
     def start(self):
-        self._ev.set()
+        self._ev.clear()
 
     def end(self):
-        self._ev.clear()
+        self._ev.set()
 
     def wait(self):
         self._ev.wait()
@@ -44,7 +45,8 @@ loader_lock = Lock()
 # These variables should only be touched while loader_lock is acquired
 swap_event = SwapEvent()
 non_local = {
-    'loaders': None
+    'loaders': None,
+    'last_action': 0.0
 }
 
 
@@ -439,21 +441,42 @@ def remove(name):
     # hitting a race condition where files are overwritten and rename operations
     # fail because the source file doesn't exist.
     if not swap_event.in_process():
-        def do_swap():
-            loader_lock.acquire()
-
-            os.remove(loader_package_path)
-            os.rename(new_loader_package_path, loader_package_path)
-
-            def do_reenable():
-                disabler.reenable_package(loader_package_name, 'loader')
-                swap_event.end()
-                loader_lock.release()
-            sublime.set_timeout(do_reenable, 10)
-
-        sublime.set_timeout(do_swap, 700)
         swap_event.start()
+        sublime.set_timeout(_do_swap, 1000)
 
+    non_local['last_action'] = time.time()
+    loader_lock.release()
+
+
+def _do_swap():
+    """
+    Swap the loader package with the new version
+    """
+
+    if not loader_lock.acquire(blocking=False):
+        sublime.set_timeout(_do_swap, 500)
+        return
+
+    if time.time() - non_local['last_action'] < 0.9:
+        sublime.set_timeout(_do_swap, 500)
+        loader_lock.release()
+        return
+
+    if os.path.exists(loader_package_path) and os.path.exists(new_loader_package_path):
+        os.remove(loader_package_path)
+        os.rename(new_loader_package_path, loader_package_path)
+
+    sublime.set_timeout(_do_reenable, 500)
+
+
+def _do_reenable():
+    """
+    Re-enable the loader package after a swap has been done
+    """
+
+    disabler = PackageDisabler()
+    disabler.reenable_package(loader_package_name, 'loader')
+    swap_event.end()
     loader_lock.release()
